+++ src/avdecc/controllerModel.cpp
@@ -28,6 +28,10 @@
 #include <algorithm>
 #include <array>
 #include <QTimer>
+#include <QFont>
+#include <set>
+#include <unordered_map>
+#include <unordered_set>
 
 Q_DECLARE_METATYPE(la::avdecc::UniqueIdentifier)
 
@@ -58,6 +62,8 @@ private:
 	Q_SLOT void controllerOffline();
 	Q_SLOT void entityOnline(la::avdecc::UniqueIdentifier const entityID);
 	Q_SLOT void entityOffline(la::avdecc::UniqueIdentifier const entityID);
+	Q_SLOT void identificationStarted(la::avdecc::UniqueIdentifier const entityID);
+	Q_SLOT void identificationStopped(la::avdecc::UniqueIdentifier const entityID);
 	Q_SLOT void entityNameChanged(la::avdecc::UniqueIdentifier const entityID, QString const& entityName);
 	Q_SLOT void entityGroupNameChanged(la::avdecc::UniqueIdentifier const entityID, QString const& entityGroupName);
 	Q_SLOT void acquireStateChanged(la::avdecc::UniqueIdentifier const entityID, la::avdecc::controller::model::AcquireState const acquireState, la::avdecc::UniqueIdentifier const owningEntity);
@@ -94,6 +100,7 @@ private:
 	using StreamsWithErrorCounter = std::set<la::avdecc::entity::model::StreamIndex>;
 	using EntitiesWithErrorCounter = std::unordered_map<la::avdecc::UniqueIdentifier, StreamsWithErrorCounter, la::avdecc::UniqueIdentifier::hash>;
 	EntitiesWithErrorCounter _entitiesWithErrorCounter{};
+	std::unordered_set<la::avdecc::UniqueIdentifier, la::avdecc::UniqueIdentifier::hash> _identifingEntities;
 
 	std::array<QImage, 5> _compatibilityImages{
 		{
@@ -122,6 +129,8 @@ ControllerModelPrivate::ControllerModelPrivate(ControllerModel* model)
 	connect(&controllerManager, &avdecc::ControllerManager::controllerOffline, this, &ControllerModelPrivate::controllerOffline);
 	connect(&controllerManager, &avdecc::ControllerManager::entityOnline, this, &ControllerModelPrivate::entityOnline);
 	connect(&controllerManager, &avdecc::ControllerManager::entityOffline, this, &ControllerModelPrivate::entityOffline);
+	connect(&controllerManager, &avdecc::ControllerManager::identificationStarted, this, &ControllerModelPrivate::identificationStarted);
+	connect(&controllerManager, &avdecc::ControllerManager::identificationStopped, this, &ControllerModelPrivate::identificationStopped);
 	connect(&controllerManager, &avdecc::ControllerManager::entityNameChanged, this, &ControllerModelPrivate::entityNameChanged);
 	connect(&controllerManager, &avdecc::ControllerManager::entityGroupNameChanged, this, &ControllerModelPrivate::entityGroupNameChanged);
 	connect(&controllerManager, &avdecc::ControllerManager::acquireStateChanged, this, &ControllerModelPrivate::acquireStateChanged);
@@ -314,6 +323,12 @@ QVariant ControllerModelPrivate::data(QModelIndex const& index, int role) const
 				}
 			}
 		}
+		else if (role == Qt::FontRole)
+		{
+			auto font = QFont{};
+			font.setBold(_identifingEntities.count(entityID) > 0);
+			return font;
+		}
 	}
 	else if (column == ControllerModel::Column::EntityLogo)
 	{
@@ -345,7 +360,7 @@ QVariant ControllerModelPrivate::data(QModelIndex const& index, int role) const
 				{
 					try
 					{
-						auto const& milanInfo = controlledEntity->getMilanInfo();
+						auto const milanInfo = *controlledEntity->getMilanInfo();
 						if ((milanInfo.featuresFlags & la::avdecc::protocol::MvuFeaturesFlags::Redundancy) == la::avdecc::protocol::MvuFeaturesFlags::Redundancy)
 						{
 							return _compatibilityImages[4];
@@ -559,6 +574,7 @@ void ControllerModelPrivate::controllerOffline()
 	q->beginResetModel();
 	_entities.clear();
 	_entitiesWithErrorCounter.clear();
+	_identifingEntities.clear();
 	q->endResetModel();
 }
 
@@ -590,8 +606,21 @@ void ControllerModelPrivate::entityOffline(la::avdecc::UniqueIdentifier const en
 		emit q->beginRemoveRows({}, row, row);
 		_entities.erase(it);
 		_entitiesWithErrorCounter.erase(entityID);
+		_identifingEntities.erase(entityID);
 		emit q->endRemoveRows();
 	}
+}
+
+void ControllerModelPrivate::identificationStarted(la::avdecc::UniqueIdentifier const entityID)
+{
+	_identifingEntities.insert(entityID);
+	dataChanged(entityID, ControllerModel::Column::EntityId, { Qt::FontRole });
+}
+
+void ControllerModelPrivate::identificationStopped(la::avdecc::UniqueIdentifier const entityID)
+{
+	_identifingEntities.erase(entityID);
+	dataChanged(entityID, ControllerModel::Column::EntityId, { Qt::FontRole });
 }
 
 void ControllerModelPrivate::entityNameChanged(la::avdecc::UniqueIdentifier const entityID, QString const& entityName)